Fix intendation errors and move to different file
